Using z component of path as scale
This allows us to zoom in and out smoothly using the path points. We
also clear away the unnecessary drawing code that is concerned with
cameras, etc.

Finally, we switch our presentation to be full screen by using the
"screen" object's width and height. This makes us dependent on our final
presentation display but hopefully this won't be a major issue.