Smoother blending from free mode to path mode
Add a "restore" mode which we enter for the transition from free mode to
path mode. During the restore we smoothstep along the line from the last
free mode position to the current path point.

We move the smoothstep function to be global so we can use it in both.

We also increase the speed.