Add Pivot point and scale to allow us to zoom into particular place
Previously zooming with the mouse would scale the image around its
original. We're interested in scaling it around the mouse position so we
record a pivot point and reference scale when pressing the RMB which we
then use during the scale calculations to scale relative to that point.

The results aren't perfect and need work.