Adjust position when starting to scale
Previously we were getting jumps when entering scale mode as the pivot
and scale were being set to something new whilst leaving the position as
it was. This new combination of values resulted in a new drawing
position.

So whilst we still want to set the pivot and scale, we also want to
adjust the drawing position so that the resulting drawing transform is
preserved given no other scaling.

We essentially calculate the combined effects of the translates and
scale in the drawing code and apply to the motion.m_pos.