Fade slides in & negate positions
As the positions are applied as 'top' and 'left' we have to negate them
when applying them if we want to be able to specify natural coordinates
in slides as the positions.

We also add CSS to fade in the current slide though I'm not sure it
works the best for going back and forth between slides but we'll see.