Use inkscape query to focus slide center
We query the groups using inkscape to get their x,y,width & height and
use that to generate a bounding box center and then a look up for the
Elm app to figure out the focus point for each slide given the name of
the slide and this lookup.

Depends on the screen resolution that maybe we should query from the
browser rather than hardcoding.